Refactor statements (and tests) into separate modules
Code hygiene only; no logical changes except rename to `AssignStatement`
